Implement General Exchange Wrappers (#106)
* Generalize exchanges

Implement exchange wrappers

compiles

respond

up

short working

* compiles

* Fixed <3

* fix

+++ contracts/lib/ContractHelper.sol
@@ -16,7 +16,7 @@ library ContractHelper {
         internal
         returns (bool)
     {
-        uint size;
+        uint256 size;
         assembly { size := extcodesize(addr) } // solium-disable-line security/no-inline-assembly
         return size > 0;
     }
